requirements.txt
numpy~=1.21.6
torch~=1.11.0+cu115
torchvision~=0.12.0+cu115
opencv-python~=4.2.0.32
scipy~=1.7.3
albumentations~=1.3.1
yaml~=0.2.5
pyyaml~=6.0
easydict~=1.11
tqdm~=4.65.0
matplotlib~=3.5.3
pillow~=9.4.0
labelme~=3.16.2
scikit-image~=0.19.3